+++ lib/Completion/Bridge/TolerantParser/WorseReflection/WorseClassMemberCompletor.php
@@ -29,6 +29,10 @@ use Phpactor\WorseReflection\Core\Reflection\ReflectionInterface;
 use Phpactor\Completion\Core\Formatter\ObjectFormatter;
 use Microsoft\PhpParser\Node\Expression\MemberAccessExpression;
 use Microsoft\PhpParser\Node\Expression\ScopedPropertyAccessExpression;
+use function in_array;
+use function str_starts_with;
+use function strlen;
+use function substr;
 
 class WorseClassMemberCompletor implements TolerantCompletor, TolerantQualifiable
 {
@@ -83,7 +87,7 @@ class WorseClassMemberCompletor implements TolerantCompletor, TolerantQualifiabl
 
         foreach ($symbolType->expandTypes()->classLike() as $type) {
             foreach ($this->populateSuggestions($nodeContext, $type, $static, $shouldCompleteOnlyName, $isInstance) as $suggestion) {
-                if ($partialMatch && 0 !== mb_strpos($suggestion->name(), $partialMatch)) {
+                if ($partialMatch && !str_starts_with($suggestion->name(), $partialMatch)) {
                     continue;
                 }