+++ lib/Completion/Tests/Integration/Bridge/TolerantParser/WorseReflection/WorseNamedParameterCompletorTest.php
@@ -109,6 +109,26 @@ class WorseNamedParameterCompletorTest extends TolerantCompletorTestCase
                 ]
             ]
         ];
+
+        yield 'Attributes' => [
+            <<<PHP
+                <?php
+                #[Attribute]
+                class SomeAttribute {
+                    public function __construct(private string \$param) {}
+                }
+
+                #[SomeAttribute(<>)]
+                class Foo {}
+                PHP,
+            [
+                [
+                    'type' => Suggestion::TYPE_FIELD,
+                    'name' => 'param: ',
+                    'short_description' => 'string $param',
+                ]
+            ]
+        ];
     }
 
     /**