+++ lib/Extension/LanguageServerCompletion/Handler/CompletionHandler.php
@@ -229,7 +229,7 @@ class CompletionHandler implements Handler, CanRegisterCapabilities
         Suggestion $suggestion,
         CompletionParams $params
     ): NameImporterResult {
-        $suggestionNameImport = $suggestion->fqn();
+        $suggestionNameImport = $suggestion->nameImport();
 
         if (!$suggestionNameImport) {
             return NameImporterResult::createEmptyResult();
@@ -279,7 +279,7 @@ class CompletionHandler implements Handler, CanRegisterCapabilities
     private function formatShortDescription(Suggestion $suggestion): string
     {
         $prefix = '';
-        if ($suggestion->classImport()) {
+        if ($suggestion->nameImport()) {
             $prefix = '↓ ';
         }