+++ lib/WorseReflection/Core/TemplateMap.php
@@ -51,11 +51,7 @@ final class TemplateMap implements Countable
         // if any of the arguments are template parameters replace them with
         // any constraints (e.g. T of Foobar)
         $arguments = array_map(function (Type $argument) {
-            if (isset($this->map[$argument->short()])) {
-                return $this->map[$argument->short()];
-            }
-
-            return $argument;
+            return $this->map[$argument->short()] ?? $argument;
         }, $arguments);
 
         if ($arguments) {